+++ README.md
@@ -33,9 +33,13 @@ To try your new server create a new pad and choose it. You can make it the defau
 The user can now share an URL to the app in order to open it. Keep in mind that the host server must be previously added, otherwise the URL will be considered unsafe. 
 
 ## Current version
-2.0
+3.0
 
-## 
+### v3.0
+- Migrated database to Room
+- Better implementation of Material Design.
+- Extreme refactoring keeping the same features and adding some minor ones.
+
 ### v2.0
 - Migrated code to Kotlin.
 - Implemented AndroidX and Material Design.
