+++ app/src/main/java/com/mikifus/padland/Dialogs/GroupPadDialog.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModelProvider
 import com.mikifus.padland.Database.PadGroupModel.PadGroup
 import com.mikifus.padland.Database.PadGroupModel.PadGroupViewModel
 import com.mikifus.padland.R
-import com.mikifus.padland.Views.Helpers.SpinnerHelper
+import com.mikifus.padland.Utils.Views.Helpers.SpinnerHelper
 
 
 /**
